refactor(mcp): keep inspection adapter within package budget
